--- src/components/create_questions/CreateExamInput.vue
@@ -1,74 +1,0 @@
-<template>
-  <label for={{ this.id }}>
-    <input type="text" id={{ this.id }} placeholder=" ">
-    <span class="label">{{ title }}</span>
-  </label>
-</template>
-
-<script>
-  export default {
-    name: 'CreateExamInput',
-    props: ['title', 'id']
-  }
-
-<style lang="scss" scoped>
-  label {
-  	width: 100%;
-    display: inline-flex;
-  }
-  .inp {
-    position: relative;
-    margin: auto;
-    width: 100%;
-  }
-  .label {
-    position: absolute;;
-    top: 13px;
-    left: 10px;
-    font-size: 16px;
-    color: #67707D;
-    transform-origin: 0 0;
-    transition: all .2s ease;
-	}
-  .border {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    height: 2px;
-    width: 100%;
-    background-color: #0077FF;
-    transform: scaleX(0);
-    transform-origin: 0 0;
-    transition: all .15s ease;
-	}
-  input {
-    -webkit-appearance: none;
-    width: 100%;
-    border: 1px solid #DBDBDB;
-		border-radius: 12px;
-    font-family: inherit;
-    padding: 7px;
-    height: 48px;
-    font-size: 16px;
-    font-weight: 500;
-    background-color: white;
-    color: #223254;
-  	transition: all .15s ease;
-	}
-  input:hover {
-      background-color: white;
-		}
-  input:not(:placeholder-shown) + span {
-  	color: #5A667F;
-    transform: translateY(-40px) translatex(-10px) scale(1);
-	}
-	input:focus {
-		outline: none;
-	}
-  input:focus + span {
-      background: none;
-      outline: none;
-      color: #67707D;
-      transform: translateY(-40px) translatex(-10px) scale(1);
-	}
-</style>
